modules/kernel: do include default initrd modules
Now that we have a full-fat kernel, all the modules NixOS expects are
present and we don't need to delete the default list. The incremental
size to the installer ISO is under 1MB so there is little downside.

In particular, this includes the kernel modules for LVM so boot stage1
doesn't have an ugly warning that device mapper communication failed,
and of course so that LVM disks can be used for boot.